refactor static super class into abstract base class - tests pass
